requirements.txt
Django==2.2.2
django-crispy-forms==1.7.2
mysqlclient==1.4.2.post1
Pillow==6.0.0
django-allauth